backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
bcrypt
python-dotenv
Flask-Migrate
psycopg2-binary
<<<<<<< HEAD

=======
prometheus-flask-exporter
>>>>>>> 45e8187e96b4861ef302eccb3bf70626916f00f2
